Move types to models
It creates a problem with Python 3.8 on docker as "types" package name
clashes with python "types"
